Vehicles related controllers and classes moved to a new package
