+++ src/main/java/com/example/vehicle/Car.java
@@ -1,4 +1,6 @@
-package com.example.rentsystem;
+package com.example.vehicle;
+
+import com.example.vehicle.Vehicle;
 
 import javax.persistence.*;
