+++ src/main/java/com/example/vehicle/VehicleController.java
@@ -1,4 +1,4 @@
-package com.example.rentsystem;
+package com.example.vehicle;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -7,7 +7,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import javax.websocket.server.PathParam;
 import java.util.Optional;
 
 @Controller // This means that this class is a Controller
