+++ src/main/java/com/example/vehicle/VehicleOwnerRepository.java
@@ -1,0 +1,7 @@
+package com.example.vehicle;
+
+import com.example.rentsystem.VehicleOwner;
+import org.springframework.data.repository.CrudRepository;
+
+public interface VehicleOwnerRepository extends CrudRepository<VehicleOwner, Integer> {
+}
